+++ src/test/java/software/amazon/encryption/s3/internal/ApiNameVersionTest.java
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package software.amazon.encryption.s3.internal;
 
-import org.junit.jupiter.api.Test;
+import org.junitpioneer.jupiter.RetryingTest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -12,7 +12,7 @@ public class ApiNameVersionTest {
     private final static String EXPECTED_API_NAME = "AmazonS3Encrypt";
     private final static String EXPECTED_API_MAJOR_VERSION = "4";
 
-    @Test
+    @RetryingTest(3)
     public void testApiNameWithVersion() {
         assertEquals(EXPECTED_API_NAME, ApiNameVersion.apiNameWithVersion().name());
         // To avoid having to hardcode versions, just check that we're incrementing from 3