+++ src/test/java/software/amazon/encryption/s3/materials/DefaultDataKeyGeneratorTest.java
@@ -3,7 +3,7 @@
 package software.amazon.encryption.s3.materials;
 
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
-import org.junit.jupiter.api.Test;
+import org.junitpioneer.jupiter.RetryingTest;
 import software.amazon.encryption.s3.algorithms.AlgorithmSuite;
 
 import javax.crypto.SecretKey;
@@ -14,7 +14,7 @@ public class DefaultDataKeyGeneratorTest {
 
     private final DataKeyGenerator dataKeyGenerator = new DefaultDataKeyGenerator();
 
-    @Test
+    @RetryingTest(3)
     public void testGenerateDataKey() {
         SecretKey actualSecretKey = dataKeyGenerator.generateDataKey(AlgorithmSuite.ALG_AES_256_GCM_IV12_TAG16_NO_KDF, null);
         assertEquals("AES", actualSecretKey.getAlgorithm());