+++ src/test/java/software/amazon/encryption/s3/materials/KmsKeyringTest.java
@@ -2,19 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0
 package software.amazon.encryption.s3.materials;
 
-import org.junit.jupiter.api.Test;
+import org.junitpioneer.jupiter.RetryingTest;
 import software.amazon.encryption.s3.S3EncryptionClientException;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class KmsKeyringTest {
 
-    @Test
+    @RetryingTest(3)
     public void buildKmsKeyringWithNullSecureRandomFails() {
         assertThrows(S3EncryptionClientException.class, () -> KmsKeyring.builder().secureRandom(null));
     }
 
-    @Test
+    @RetryingTest(3)
     public void buildKmsKeyringWithNullDataKeyGeneratorFails() {
         assertThrows(S3EncryptionClientException.class, () -> KmsKeyring.builder().dataKeyGenerator(null));
     }